backend/signal_chat.py
"""
Signal Protocol E2EE Chat Implementation
Double-ratchet encryption for medical communications
Zero-knowledge server: cannot read messages
"""

from axolotl.state.prekeybundle import PreKeyBundle
from axolotl.sessionbuilder import SessionBuilder
from axolotl.sessioncipher import SessionCipher
from axolotl.protocol.prekeywhispermessage import PreKeyWhisperMessage
from axolotl.protocol.whispermessage import WhisperMessage
from axolotl.identitykey import IdentityKey
from axolotl.state.axolotlstore import AxolotlStore
from axolotl.util.keyhelper import KeyHelper

import sqlite3
import json
import base64
from datetime import datetime
from typing import Optional, List, Dict, Any
import logging

logger = logging.getLogger(__name__)


class SignalProtocolStore(AxolotlStore):
    """
    Signal Protocol key storage
    Implements AxolotlStore interface for key management
    """
    
    def __init__(self, user_id: str, db_path: str = "chat_keys.db"):
        self.user_id = user_id
        self.db_path = db_path
        self.init_database()
        
        # Load or generate identity keys
        identity_keypair = self.load_identity_keypair()
        if not identity_keypair:
            identity_keypair = KeyHelper.generateIdentityKeyPair()
            self.save_identity_keypair(identity_keypair)
        
        self.identityKeyPair = identity_keypair
        self.localRegistrationId = KeyHelper.generateRegistrationId()
    
    def init_database(self):
        """Initialize key storage database"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        # Identity keys
        cursor.execute("""
            CREATE TABLE IF NOT EXISTS identity_keys (
                user_id TEXT PRIMARY KEY,
                public_key TEXT,
                private_key TEXT
            )
        """)
        
        # Pre-keys
        cursor.execute("""
            CREATE TABLE IF NOT EXISTS prekeys (
                user_id TEXT,
                key_id INTEGER,
                key_pair TEXT,
                PRIMARY KEY (user_id, key_id)
            )
        """)
        
        # Sessions
        cursor.execute("""
            CREATE TABLE IF NOT EXISTS sessions (
                user_id TEXT,
                recipient_id TEXT,
                device_id INTEGER,
                session_record TEXT,
                PRIMARY KEY (user_id, recipient_id, device_id)
            )
        """)
        
        conn.commit()
        conn.close()
    
    def load_identity_keypair(self):
        """Load identity key pair"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        cursor.execute("""
            SELECT public_key, private_key FROM identity_keys WHERE user_id = ?
        """, (self.user_id,))
        
        row = cursor.fetchone()
        conn.close()
        
        if row:
            # Reconstruct key pair from stored data
            # (Implementation depends on axolotl library format)
            return None  # Simplified for now
        return None
    
    def save_identity_keypair(self, keypair):
        """Save identity key pair"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        # Simplified storage (in production, properly serialize keys)
        cursor.execute("""
            INSERT OR REPLACE INTO identity_keys (user_id, public_key, private_key)
            VALUES (?, ?, ?)
        """, (self.user_id, str(keypair.getPublicKey()), "encrypted_private"))
        
        conn.commit()
        conn.close()


class SecureChatManager:
    """
    Manages Signal Protocol encrypted chat
    Server sees only: sender, recipient, timestamp, ciphertext
    """
    
    def __init__(self, db_path: str = "chat_messages.db"):
        self.db_path = db_path
        self.init_database()
        self.active_sessions = {}  # In-memory session cache
    
    def init_database(self):
        """Initialize chat database"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        # Users table
        cursor.execute("""
            CREATE TABLE IF NOT EXISTS users (
                user_id TEXT PRIMARY KEY,
                username TEXT UNIQUE,
                public_identity_key TEXT,
                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
            )
        """)
        
        # Encrypted messages (server stores ONLY ciphertext)
        cursor.execute("""
            CREATE TABLE IF NOT EXISTS messages (
                message_id TEXT PRIMARY KEY,
                sender_id TEXT,
                recipient_id TEXT,
                ciphertext BLOB,
                timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                is_group BOOLEAN DEFAULT 0,
                group_id TEXT,
                FOREIGN KEY (sender_id) REFERENCES users(user_id),
                FOREIGN KEY (recipient_id) REFERENCES users(user_id)
            )
        """)
        
        # Groups
        cursor.execute("""
            CREATE TABLE IF NOT EXISTS chat_groups (
                group_id TEXT PRIMARY KEY,
                group_name TEXT,
                created_by TEXT,
                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
            )
        """)
        
        # Group members
        cursor.execute("""
            CREATE TABLE IF NOT EXISTS group_members (
                group_id TEXT,
                user_id TEXT,
                joined_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                PRIMARY KEY (group_id, user_id),
                FOREIGN KEY (group_id) REFERENCES chat_groups(group_id),
                FOREIGN KEY (user_id) REFERENCES users(user_id)
            )
        """)
        
        conn.commit()
        conn.close()
        logger.info("Chat database initialized")
    
    def register_user(self, user_id: str, username: str, public_key: str):
        """Register user with their public identity key"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        cursor.execute("""
            INSERT OR REPLACE INTO users (user_id, username, public_identity_key)
            VALUES (?, ?, ?)
        """, (user_id, username, public_key))
        
        conn.commit()
        conn.close()
        logger.info(f"User registered: {username}")
    
    def send_encrypted_message(self, sender_id: str, recipient_id: str, 
                              ciphertext: bytes, group_id: Optional[str] = None):
        """
        Store encrypted message
        Server CANNOT read the message content
        """
        import uuid
        message_id = str(uuid.uuid4())
        
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        is_group = group_id is not None
        
        cursor.execute("""
            INSERT INTO messages (message_id, sender_id, recipient_id, ciphertext, is_group, group_id)
            VALUES (?, ?, ?, ?, ?, ?)
        """, (message_id, sender_id, recipient_id, ciphertext, is_group, group_id))
        
        conn.commit()
        conn.close()
        
        logger.info(f"Encrypted message stored: {message_id} (server cannot read)")
        return message_id
    
    def get_messages(self, user_id: str, limit: int = 50) -> List[Dict[str, Any]]:
        """
        Get encrypted messages for user
        Returns ciphertext only - client must decrypt
        """
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        cursor.execute("""
            SELECT message_id, sender_id, recipient_id, ciphertext, timestamp, is_group, group_id
            FROM messages
            WHERE recipient_id = ? OR sender_id = ?
            ORDER BY timestamp DESC
            LIMIT ?
        """, (user_id, user_id, limit))
        
        rows = cursor.fetchall()
        conn.close()
        
        return [{
            'message_id': row[0],
            'sender_id': row[1],
            'recipient_id': row[2],
            'ciphertext': base64.b64encode(row[3]).decode(),  # Server sends encrypted blob
            'timestamp': row[4],
            'is_group': bool(row[5]),
            'group_id': row[6]
        } for row in rows]
    
    def create_group(self, group_id: str, group_name: str, created_by: str, member_ids: List[str]):
        """Create encrypted group chat"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        # Create group
        cursor.execute("""
            INSERT INTO chat_groups (group_id, group_name, created_by)
            VALUES (?, ?, ?)
        """, (group_id, group_name, created_by))
        
        # Add members
        for user_id in member_ids:
            cursor.execute("""
                INSERT INTO group_members (group_id, user_id)
                VALUES (?, ?)
            """, (group_id, user_id))
        
        conn.commit()
        conn.close()
        logger.info(f"Group created: {group_name} with {len(member_ids)} members")
    
    def get_user_groups(self, user_id: str) -> List[Dict[str, Any]]:
        """Get groups user belongs to"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        cursor.execute("""
            SELECT g.group_id, g.group_name, g.created_by, g.created_at
            FROM chat_groups g
            JOIN group_members gm ON g.group_id = gm.group_id
            WHERE gm.user_id = ?
        """, (user_id,))
        
        rows = cursor.fetchall()
        conn.close()
        
        return [{
            'group_id': row[0],
            'group_name': row[1],
            'created_by': row[2],
            'created_at': row[3]
        } for row in rows]
    
    def delete_message(self, message_id: str, user_id: str):
        """Delete message (only if sender)"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        cursor.execute("""
            DELETE FROM messages 
            WHERE message_id = ? AND sender_id = ?
        """, (message_id, user_id))
        
        conn.commit()
        deleted = cursor.rowcount
        conn.close()
        
        return deleted > 0


# Simple client-side encryption helper (for demo)
class SimpleE2EEClient:
    """
    Simplified E2EE client for demonstration
    In production, use full Signal Protocol implementation
    """
    
    def __init__(self, user_id: str, encryption_key: bytes):
        self.user_id = user_id
        self.key = encryption_key
    
    def encrypt_message(self, message: str) -> bytes:
        """Encrypt message (simplified AES-256-GCM for demo)"""
        from cryptography.hazmat.primitives.ciphers.aead import AESGCM
        import os
        
        nonce = os.urandom(12)
        aesgcm = AESGCM(self.key)
        ciphertext = aesgcm.encrypt(nonce, message.encode(), None)
        
        return nonce + ciphertext
    
    def decrypt_message(self, ciphertext: bytes) -> str:
        """Decrypt message"""
        from cryptography.hazmat.primitives.ciphers.aead import AESGCM
        
        nonce = ciphertext[:12]
        encrypted = ciphertext[12:]
        
        aesgcm = AESGCM(self.key)
        plaintext = aesgcm.decrypt(nonce, encrypted, None)
        
        return plaintext.decode()
