backend/requirements.txt
# Medical Document Summarizer - Requirements

# FastAPI Backend
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.12.0

# Streamlit Frontend
streamlit==1.53.1
pandas==3.0.0

# AI/ML Models
torch==2.5.1
transformers==5.0.0
accelerate==1.12.0
sentencepiece==0.2.1
safetensors==0.7.0
bitsandbytes==0.49.1

# NER & NLP
spacy==3.8.11
# Run after install: python -m spacy download en_core_web_sm

# OCR
pytesseract==0.3.13
# Requires Tesseract binary: https://github.com/UB-Mannheim/tesseract/wiki

# Document Processing
PyPDF2==3.0.1
python-docx==1.2.0
pdf2image==1.17.0
pillow==12.1.0

# LangChain & LangGraph
langgraph==1.0.7
langchain==1.2.7
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0

# Encryption & Security
cryptography==46.0.4
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
protobuf==3.20.3
PyNaCl==1.6.2

# Database
SQLAlchemy==2.0.46

# Audio/TTS
pyttsx3==2.99
pygame==2.6.1

# WebSockets
websockets==16.0

# HTTP Clients
requests==2.32.5
httpx==0.28.1
aiohttp==3.13.3

# Utilities
python-dotenv==1.2.1
PyYAML==6.0.3
tqdm==4.67.2
